+++ CHANGELOG.md
@@ -3,12 +3,20 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htmlhttps://semver.org/spec/v2.0.0.html).
 
-## 0.2.5
+## Upcoming
+### Added
+* Warning box if composite data does not exist
+* Scaling for single color images
+* Undo of thematic map edits
+
+## [0.2.5]
 ### Added
 * Left click on the thematic map relabels the contiguous region
 * Right click on the thematic map draws boundaries on the preview image
 
-## 0.2.4
-First stable release.
+## [0.2.4]
+First stable release. Not all features added. 
+
+[0.2.5]: https://github.com/jmbhughes/solarannotator/releases/tag/v0.2.5 
+[0.2.4]: https://github.com/jmbhughes/solarannotator/releases/tag/v0.2.4